split streamers
